day 11 pt 2

+++ src/main/kotlin/day11/Coord.kt
@@ -1,3 +1,3 @@
 package day11
 
-data class Coord(val row: Int, val col: Int)
+data class Coord(val row: Long, val col: Long)
